+++ src/main/kotlin/com/pswidersk/gradle/python/EnvSetupTask.kt
@@ -1,7 +1,6 @@
 package com.pswidersk.gradle.python
 
 import org.gradle.api.DefaultTask
-import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 import org.gradle.process.ExecOperations
 import org.gradle.process.ExecResult
@@ -11,8 +10,7 @@ abstract class EnvSetupTask @Inject constructor(
     private val execOperations: ExecOperations,
 ) : DefaultTask() {
 
-    @Internal
-    val pythonPluginExtension: PythonPluginExtension = project.pythonPlugin
+    private val pythonPluginExtension: PythonPluginExtension = project.pythonPlugin
 
     init {
         group = "python"
@@ -23,7 +21,7 @@ abstract class EnvSetupTask @Inject constructor(
     @TaskAction
     fun setup(): ExecResult = with(pythonPluginExtension) {
         execOperations.exec {
-            it.executable = condaExec.get().asFile.absolutePath
+            it.executable = condaExec.get().asFile.canonicalPath
             it.args(listOf("create", "--name", pythonEnvName.get(), "python=${pythonVersion.get()}"))
         }
     }