+++ src/main/kotlin/com/pswidersk/gradle/python/ListPropertiesTask.kt
@@ -1,13 +1,11 @@
 package com.pswidersk.gradle.python
 
 import org.gradle.api.DefaultTask
-import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 
 abstract class ListPropertiesTask : DefaultTask() {
 
-    @Internal
-    val pythonPluginExtension: PythonPluginExtension = project.pythonPlugin
+    private val pythonPluginExtension: PythonPluginExtension = project.pythonPlugin
 
     init {
         group = "python"