+++ src/main/kotlin/com/pswidersk/gradle/python/PythonPluginExtension.kt
@@ -8,6 +8,7 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.ProviderFactory
 import org.gradle.internal.os.OperatingSystem
+import org.gradle.kotlin.dsl.property
 import javax.inject.Inject
 
 abstract class PythonPluginExtension @Inject constructor(