+++ src/main/kotlin/com/pswidersk/gradle/python/VenvTask.kt
@@ -4,16 +4,14 @@ import org.apache.tools.ant.types.Commandline
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.ProjectLayout
+import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.options.Option
 import org.gradle.process.ExecOperations
 import org.gradle.process.ExecResult
 import org.gradle.process.internal.streams.SafeStreams
-import java.io.InputStream
-import java.io.OutputStream
 import javax.inject.Inject
 
 abstract class VenvTask @Inject constructor(
@@ -22,13 +20,11 @@ abstract class VenvTask @Inject constructor(
     projectLayout: ProjectLayout
 ) : DefaultTask() {
 
-    @Internal
-    val pythonPluginExtension: PythonPluginExtension = project.pythonPlugin
+    private val pythonPluginExtension: PythonPluginExtension = project.pythonPlugin
 
     init {
         group = PLUGIN_TASKS_GROUP_NAME
         this.dependsOn("envSetup")
-        notCompatibleWithConfigurationCache("Input and Output streams are disallowed fields for config cache.")
     }
 
     /**
@@ -36,24 +32,42 @@ abstract class VenvTask @Inject constructor(
      * For example: "install", "./main.py"
      *
      */
-    @Input
+    @Internal
     var args: List<String> = emptyList()
 
     /**
      * Working directory
      *
      */
-    @Input
+    @Internal
     val workingDir: DirectoryProperty = objects.directoryProperty().convention(projectLayout.projectDirectory)
 
-    @Input
+    /**
+     * Environment variables map
+     */
+    @Internal
     var environment: Map<String, Any> = mapOf()
 
-    @Input
-    var standardInput: InputStream = SafeStreams.emptyInput()
+    /**
+     * Input file to be passed into standard input.
+     */
+    @Internal
+    val inputFile: RegularFileProperty = objects.fileProperty()
 
-    @Input
-    var standardOutput: OutputStream = SafeStreams.systemOut()
+    /**
+     * Output file to be passed into standard output.
+     */
+    @Internal
+    val outputFile: RegularFileProperty = objects.fileProperty()
+
+    /**
+     * Executable which have to exist in virtual env.
+     * For example: "python", "pip", "wheel"
+     *
+     * Default: "python"
+     */
+    @Internal
+    var venvExec: String = "python"
 
     /**
      * Parses an argument list from {@code args} and passes it to args.
@@ -76,15 +90,6 @@ abstract class VenvTask @Inject constructor(
     fun setArgsByCmd(args: String) {
         this.args = Commandline.translateCommandline(args).toList()
     }
-
-    /**
-     * Executable which have to exist in virtual env.
-     * For example: "python", "pip", "wheel"
-     *
-     * Default: "python"
-     */
-    @Input
-    var venvExec: String = "python"
 
     @TaskAction
     fun execute(): ExecResult = with(pythonPluginExtension) {
@@ -110,8 +115,10 @@ abstract class VenvTask @Inject constructor(
             it.commandLine(allArgs)
             it.workingDir(workingDir)
             it.environment(environment)
-            it.standardInput = standardInput
-            it.standardOutput = standardOutput
+            it.standardInput =
+                if (inputFile.isPresent) inputFile.get().asFile.inputStream() else SafeStreams.emptyInput()
+            it.standardOutput =
+                if (outputFile.isPresent) outputFile.get().asFile.outputStream() else SafeStreams.systemOut()
         }
     }
 }