+++ src/test/kotlin/com/pswidersk/gradle/python/PythonPluginTest.kt
@@ -2,6 +2,7 @@ package com.pswidersk.gradle.python
 
 import org.assertj.core.api.Assertions.assertThat
 import org.gradle.api.Project
+import org.gradle.kotlin.dsl.apply
 import org.gradle.testfixtures.ProjectBuilder
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
@@ -17,7 +18,7 @@ internal class PythonPluginTest {
     @Test
     fun `test if plugin was applied`() {
         val project: Project = ProjectBuilder.builder().build()
-        project.pluginManager.apply(PythonPlugin::class.java)
+        project.apply<PythonPlugin>()
 
         assertThat(project.plugins.size).isEqualTo(1)
     }
@@ -43,7 +44,7 @@ internal class PythonPluginTest {
             .withPluginClasspath()
             .withProjectDir(tempDir)
             .forwardOutput()
-            .withArguments(":listPluginProperties")
+            .withArguments("--configuration-cache", ":listPluginProperties")
 
         // when
         val runResult = runner.build()
@@ -141,7 +142,7 @@ internal class PythonPluginTest {
             .withPluginClasspath()
             .withProjectDir(tempDir)
             .forwardOutput()
-            .withArguments("--configuration-cache", ":runTestScript")
+            .withArguments("--configuration-cache", "--warning-mode", "all", "--info", ":runTestScript")
 
         // when
         val firstRunResult = runner.build()